pyproject.toml
[project]
name = "squirrel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.30.1",
    "ipython-sql>=0.5.0",
    "mysqlclient>=2.2.7",
    "pandas>=2.3.1",
    "prettytable==3.11",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.5",
]

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "single"
